Use OCMock to eliminate testWithoutNetwork change in library
