+++ FirebaseRemoteConfig.podspec
@@ -74,20 +74,21 @@ app update.
   s.test_spec 'swift-api-tests' do |swift_api|
     swift_api.platforms = {:ios => '8.0', :osx => '10.11', :tvos => '10.0'}
     swift_api.source_files = 'FirebaseRemoteConfig/Tests/SwiftAPI/*.swift',
-                             'FirebaseRemoteConfig/Tests/FakeUtils/*.h',
+                             'FirebaseRemoteConfig/Tests/FakeUtils/*.[hm]',
                              'FirebaseRemoteConfig/Tests/FakeUtils/*.swift'
     swift_api.requires_app_host = true
     swift_api.pod_target_xcconfig = {
       'SWIFT_OBJC_BRIDGING_HEADER' => '$(PODS_TARGET_SRCROOT)/FirebaseRemoteConfig/Tests/FakeUtils/Bridging-Header.h'
     }
     swift_api.resources = 'FirebaseRemoteConfig/Tests/SwiftAPI/GoogleService-Info.plist'
+    swift_api.dependency 'OCMock'
   end
 
   # Run Swift API tests and tests requiring console changes on a Fake Console.
   s.test_spec 'fake-console-tests' do |fake_console|
     fake_console.platforms = {:ios => '8.0', :osx => '10.11', :tvos => '10.0'}
     fake_console.source_files = 'FirebaseRemoteConfig/Tests/SwiftAPI/*.swift',
-                                      'FirebaseRemoteConfig/Tests/FakeUtils/*.h',
+                                      'FirebaseRemoteConfig/Tests/FakeUtils/*.[hm]',
                                       'FirebaseRemoteConfig/Tests/FakeUtils/*.swift',
                                       'FirebaseRemoteConfig/Tests/FakeConsole/*.swift'
     fake_console.requires_app_host = true
@@ -95,5 +96,6 @@ app update.
       'SWIFT_OBJC_BRIDGING_HEADER' => '$(PODS_TARGET_SRCROOT)/FirebaseRemoteConfig/Tests/FakeUtils/Bridging-Header.h'
     }
     fake_console.resources = 'FirebaseRemoteConfig/Tests/FakeUtils/GoogleService-Info.plist'
+    fake_console.dependency 'OCMock'
   end
 end
