+++ FirebaseRemoteConfig/Tests/FakeConsole/FakeConsoleTests.swift
@@ -17,39 +17,11 @@ import FirebaseCore
 
 import XCTest
 
-class FakeConsoleTests: XCTestCase {
-  static var useFakeConfig: Bool!
-  var app: FirebaseApp!
-  var config: RemoteConfig!
-  var fakeConsole: FakeConsole!
-
-  override class func setUp() {
-    if !(FirebaseApp.app() != nil) {
-      FirebaseApp.configure()
-    }
-    useFakeConfig = FirebaseApp.app()!.options.projectID == "FakeProject"
-  }
+class FakeConsoleTests: APITestBase {
 
   override func setUp() {
     super.setUp()
-    app = FirebaseApp.app()
-    config = RemoteConfig.remoteConfig(app: app!)
-    let settings = RemoteConfigSettings()
-    settings.minimumFetchInterval = 0
-    config.configSettings = settings
-    fakeConsole = FakeConsole(with: ["Key1": "Value1"])
-    config.configFetch.fetchSession = URLSessionMock(with: fakeConsole)
-    config.configFetch.testWithoutNetwork = true
-
-    // Uncomment for verbose debug logging.
-    // FirebaseConfiguration.shared.setLoggerLevel(FirebaseLoggerLevel.debug)
-  }
-
-  override func tearDown() {
-    app = nil
-    config = nil
-    fakeConsole.empty()
-    super.tearDown()
+    fakeConsole.config = ["Key1": "Value1"]
   }
 
   // Contrast with testUnchangedActivateWillError in APITests.swift.
